v2/requirements.txt
requests-kerberos==0.15.0
pykerberos==1.2.4
